Scheduler added, postRepository DTO added

+++ src/main/java/diplom/blogengine/api/response/PostResponse.java
@@ -17,7 +17,7 @@ public class PostResponse {
     private long likeCount;
     private long dislikeCount;
     private long commentCount;
-    private long viewCount;
+    private int viewCount;
 
     @JsonPOJOBuilder(withPrefix = "")
     public static class PostResponseBuilder {
